refactor(types): replace type aliases with direct class definitions in `.pyi` files
Refactors type stub files to use direct class definitions instead of type aliases, aligning with updated implementation practices. Updates multiple modules for consistency and enhanced clarity.
